+++ src/App.tsx
@@ -496,6 +496,7 @@ class App extends React.Component <{
 
 	async getClasses(): Promise<void> {
 		this.setState({loading: true})
+		// TODO: Send message when there are no selected Degrees
 		const [classesByShift, minimalClasses] = await getMinimalClasses(this.state.selectedShifts, this.selectedDegrees)
 
 		this.classesByShift = Object.entries(classesByShift)